% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs_for_app.R
\name{clean.table}
\alias{clean.table}
\title{select, rename columns and define some format based on output of adhoc region
by default add Sex = "Total"}
\usage{
clean.table(dt)
}
\description{
select, rename columns and define some format based on output of adhoc region
by default add Sex = "Total"
}
