% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs_for_app.R
\name{read.country.summary}
\alias{read.country.summary}
\title{function to read "Rates & Deaths_Country Summary.csv" and output long format
\code{OfficialName} column named as \code{Region} for easier binding later in the
downloaded file}
\usage{
read.country.summary(dir_dt_cs, year_wanted = NULL, sex = NULL)
}
\description{
function to read "Rates & Deaths_Country Summary.csv" and output long format
\code{OfficialName} column named as \code{Region} for easier binding later in the
downloaded file
}
