% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs_for_app.R
\name{read.region.summary}
\alias{read.region.summary}
\title{function to read "Rates & Deaths_(region name).csv" and output long format
data}
\usage{
read.region.summary(
  dir_file,
  year_range = NULL,
  sex = NULL,
  add_regional_grouping = FALSE
)
}
\description{
function to read "Rates & Deaths_(region name).csv" and output long format
data
}
